move sanity check functions to l2_relayer_sanity.go
